Moved process rekognition results to separate file.
